writer examples
