+++ topic/topicwriter/topicwriter.go
@@ -1,0 +1,52 @@
+package topicwriter
+
+import (
+	"bytes"
+	"context"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3"
+	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topicoptions"
+	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topictypes"
+	"github.com/ydb-platform/ydb-go-sdk/v3/topic/topicwriter"
+)
+
+func ConnectSimple(ctx context.Context, db ydb.Connection) *topicwriter.Writer {
+	producerAndGroupID := "group-id"
+	writer, _ := db.Topic().StartWriter(producerAndGroupID, "topicName",
+		topicoptions.WithMessageGroupID(producerAndGroupID),
+	)
+	return writer
+}
+
+func ConnectWithSyncWrite(ctx context.Context, db ydb.Connection) *topicwriter.Writer {
+	producerAndGroupID := "group-id"
+	writer, _ := db.Topic().StartWriter(producerAndGroupID, "topicName",
+		topicoptions.WithMessageGroupID(producerAndGroupID),
+		topicoptions.WithSyncWrite(true),
+	)
+	return writer
+}
+
+func ConnectSelectCodec(ctx context.Context, db ydb.Connection) *topicwriter.Writer {
+	producerAndGroupID := "group-id"
+	writer, _ := db.Topic().StartWriter(producerAndGroupID, "topicName",
+		topicoptions.WithMessageGroupID(producerAndGroupID),
+		topicoptions.WithCodec(topictypes.CodecGzip),
+	)
+	return writer
+}
+
+func SendMessagesOneByOne(ctx context.Context, w *topicwriter.Writer) {
+	data := []byte{1, 2, 3}
+	mess := topicwriter.Message{Data: bytes.NewReader(data)}
+	_ = w.Write(ctx, mess)
+}
+
+func SendGroupOfMessages(ctx context.Context, w *topicwriter.Writer) {
+	data1 := []byte{1, 2, 3}
+	data2 := []byte{4, 5, 6}
+	mess1 := topicwriter.Message{Data: bytes.NewReader(data1)}
+	mess2 := topicwriter.Message{Data: bytes.NewReader(data2)}
+
+	_ = w.Write(ctx, mess1, mess2)
+}
